Refactor payment-provisioning component into core-structure
