+++ packages/webcomponents/src/components/business-forms/payment-provisioning/payment-provisioning.tsx
@@ -1,13 +1,10 @@
-import { Component, Host, h, Prop, State, Event, EventEmitter } from '@stencil/core';
-import { LoadingSpinner } from '../../form/utils';
-import { BusinessFormClickActions, BusinessFormClickEvent, BusinessFormSubmitEvent } from '../utils/business-form-types';
+import { Component, h, Prop, State, Event, EventEmitter, Watch } from '@stencil/core';
 import JustifiAnalytics from '../../../api/Analytics';
 import { ComponentError, ComponentErrorCodes, ComponentErrorSeverity } from '../../../api/ComponentError';
 import { makeGetBusiness } from './business-actions';
 import { makePostProvisioning } from './provision-actions';
 import { BusinessService } from '../../../api/services/business.service';
 import { ProvisionService } from '../../../api/services/provision.service';
-import { checkProvisioningStatus } from '../utils/helpers';
 
 /**
  * @exportedPart label: Label for inputs
@@ -21,10 +18,6 @@ import { checkProvisioningStatus } from '../utils/helpers';
   shadow: true
 })
 export class PaymentProvisioning {
-  @State() formLoading: boolean = false;
-  @State() businessProvisioned: boolean = false;
-  @State() currentStep: number = 0;
-  @State() errorMessage: string;
   @State() postProvisioning: Function;
   @State() getBusiness: Function;
 
@@ -32,20 +25,20 @@ export class PaymentProvisioning {
   @Prop() authToken: string;
   @Prop() allowOptionalFields?: boolean = false;
   @Prop() formTitle?: string = 'Business Information';
-  @Prop() removeTitle?: boolean = false;
 
-  @Event({ eventName: 'click-event' }) clickEvent: EventEmitter<BusinessFormClickEvent>;
+  @Watch('authToken')
+  @Watch('businessId')
+  propChanged() {
+    this.initializeApi();
+  }
+
   @Event({ eventName: 'error-event' }) errorEvent: EventEmitter<ComponentError>;
-  @Event() submitted: EventEmitter<BusinessFormSubmitEvent>;
 
   analytics: JustifiAnalytics;
 
   componentWillLoad() {
     this.analytics = new JustifiAnalytics(this);
     this.initializeApi();
-    this.fetchBusiness();
-
-    this.refs = [this.coreInfoRef, this.legalAddressRef, this.additionalQuestionsRef, this.representativeRef, this.ownersRef, this.bankAccountRef, this.documentUploadRef, this.termsRef];
   }
 
   disconnectedCallback() {
@@ -74,214 +67,16 @@ export class PaymentProvisioning {
     }
   }
 
-  fetchBusiness() {
-    this.getBusiness({
-      onSuccess: (response) => {
-        this.businessProvisioned = checkProvisioningStatus(response.data);
-        if (this.businessProvisioned) {
-          this.errorEvent.emit({
-            message: 'A request to provision payments for this business has already been submitted.',
-            errorCode: ComponentErrorCodes.PROVISIONING_REQUESTED,
-            severity: ComponentErrorSeverity.INFO,
-          });
-        }
-      },
-      onError: ({ error, code, severity }) => {
-        this.errorEvent.emit({
-          message: error,
-          errorCode: code,
-          severity: severity
-        });
-      }
-    });
-  }
-
-  postProvisioningData() {
-    this.postProvisioning({
-      onSuccess: (response) => {
-        this.submitted.emit({ data: { response } });
-      },
-      onError: ({ error, code, severity }) => {
-        this.submitted.emit({ data: { error } });
-        this.errorEvent.emit({
-          message: error,
-          errorCode: code,
-          severity: severity
-        });
-      }
-    });
-  }
-  
-  get title() {
-    return this.removeTitle ? '' : this.formTitle;
-  }
-
-  get totalSteps() {
-    return Object.keys(this.componentStepMapping).length - 1;
-  }
-
-  get businessEndpoint() {
-    return `entities/business/${this.businessId}`
-  }
-
-  get formDisabled() {
-    return this.formLoading || this.businessProvisioned;
-  }
-
-  private coreInfoRef: any;
-  private legalAddressRef: any;
-  private additionalQuestionsRef: any;
-  private representativeRef: any;
-  private ownersRef: any;
-  private bankAccountRef: any;
-  private documentUploadRef: any;
-  private termsRef: any;
-  private refs = [];
-
-  componentStepMapping = {
-    0: () => <justifi-business-core-info-form-step
-      businessId={this.businessId}
-      authToken={this.authToken}
-      ref={(el) => this.refs[0] = el}
-      onFormLoading={this.handleFormLoading}
-      allowOptionalFields={this.allowOptionalFields}
-    />,
-    1: () => <justifi-legal-address-form-step
-      businessId={this.businessId}
-      authToken={this.authToken}
-      ref={(el) => this.refs[1] = el}
-      onFormLoading={this.handleFormLoading}
-      allowOptionalFields={this.allowOptionalFields}
-    />,
-    2: () => <justifi-additional-questions-form-step
-      businessId={this.businessId}
-      authToken={this.authToken}
-      ref={(el) => this.refs[2] = el}
-      onFormLoading={this.handleFormLoading}
-      allowOptionalFields={this.allowOptionalFields}
-    />,
-    3: () => <justifi-business-representative-form-step
-      businessId={this.businessId}
-      authToken={this.authToken}
-      ref={(el) => this.refs[3] = el}
-      onFormLoading={this.handleFormLoading}
-      allowOptionalFields={this.allowOptionalFields}
-    />,
-    4: () => <justifi-business-owners-form-step
-      businessId={this.businessId}
-      authToken={this.authToken}
-      ref={(el) => this.refs[4] = el}
-      onFormLoading={this.handleFormLoading}
-      allowOptionalFields={this.allowOptionalFields}
-    />,
-    5: () => <justifi-business-bank-account-form-step
-      businessId={this.businessId}
-      authToken={this.authToken}
-      ref={(el) => this.refs[5] = el}
-      onFormLoading={this.handleFormLoading}
-      allowOptionalFields={this.allowOptionalFields}
-    />,
-    6: () => <justifi-business-document-upload-form-step
-      businessId={this.businessId}
-      authToken={this.authToken}
-      ref={(el) => this.refs[6] = el}
-      onFormLoading={this.handleFormLoading}
-      allowOptionalFields={this.allowOptionalFields}
-    />,
-    7: () => <justifi-business-terms-conditions-form-step
-      businessId={this.businessId}
-      authToken={this.authToken}
-      ref={(el) => this.refs[7] = el}
-      onFormLoading={this.handleFormLoading}
-      allowOptionalFields={this.allowOptionalFields}
-      onSubmitted={() => this.postProvisioningData()}
-    />,
-  };
-
-  handleFormLoading = (e: CustomEvent) => {
-    this.formLoading = e.detail;
-  }
-
-  showPreviousStepButton() {
-    return this.currentStep > 0;
-  }
-
-  showNextStepButton() {
-    return this.currentStep < this.totalSteps;
-  }
-
-  incrementSteps = () => {
-    if (this.currentStep < this.totalSteps) {
-      return this.currentStep++;
-    }
-  }
-  decrementSteps = () => { return this.currentStep--; }
-
-  previousStepButtonOnClick() {
-    this.clickEvent.emit({ name: BusinessFormClickActions.previousStep })
-    this.decrementSteps();
-  }
-
-  nextStepButtonOnClick(e: any, clickEventName) {
-    e.preventDefault();
-    this.clickEvent.emit({ name: clickEventName })
-
-    const currentStep = this.refs[this.currentStep];
-    currentStep.validateAndSubmit({ onSuccess: this.incrementSteps });
-  }
-
-  showSubmitButton() {
-    return this.currentStep === this.totalSteps;
-  }
-
-  currentStepComponent() {
-    return this.componentStepMapping[this.currentStep]();
-  }
-
   render() {
     return (
-      <Host exportparts="label,input,input-invalid">
-        <div class="row gap-3">
-          <h1>{this.title}</h1>
-          <div class="col-12 mb-4">
-            {this.currentStepComponent()}
-          </div>
-          <div class="d-flex justify-content-between align-items-center">
-            <div class="d-flex align-items-center">
-              Step {this.currentStep + 1} of {this.totalSteps + 1}
-            </div>
-            <div class="d-flex gap-2">
-              {this.showPreviousStepButton() && (
-                <button
-                  type="button"
-                  class="btn btn-secondary"
-                  onClick={() => this.previousStepButtonOnClick()}
-                  disabled={this.formDisabled}>
-                  Previous
-                </button>
-              )}
-              {this.showNextStepButton() && (
-                <button
-                  type="button"
-                  class={`btn btn-primary jfi-submit-button${this.formLoading ? ' jfi-submit-button-loading' : ''}`}
-                  onClick={(e) => this.nextStepButtonOnClick(e, BusinessFormClickActions.nextStep)}
-                  disabled={this.formDisabled}>
-                  {this.formLoading ? LoadingSpinner() : 'Next'}
-                </button>
-              )}
-              {this.showSubmitButton() && (
-                <button
-                  type="submit"
-                  class={`btn btn-primary jfi-submit-button${this.formLoading ? ' jfi-submit-button-loading' : ''}`}
-                  onClick={(e) => this.nextStepButtonOnClick(e, BusinessFormClickActions.submit)}
-                  disabled={this.formDisabled}>
-                  {this.formLoading ? LoadingSpinner() : 'Submit'}
-                </button>
-              )}
-            </div>
-          </div>
-        </div>
-      </Host>
+      <justifi-payment-provisioning-core
+        businessId={this.businessId}
+        authToken={this.authToken}
+        allowOptionalFields={this.allowOptionalFields}
+        formTitle={this.formTitle}
+        getBusiness={this.getBusiness}
+        postProvisioning={this.postProvisioning}
+      />
     );
   }
 }
